+++ docs/components/Combobox/Web.stories.tsx
@@ -1,7 +1,8 @@
 import React, { useState } from "react";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { Combobox, ComboboxOption } from "@jobber/components/Combobox";
-import { Button as ClearButton } from "@jobber/components/Button";
+import { Button, Button as ClearButton } from "@jobber/components/Button";
+import { Chip } from "@jobber/components/Chip";
 
 export default {
   title: "Components/Selections/Combobox/Web",
@@ -10,20 +11,61 @@ export default {
     viewMode: "story",
     previewTabs: { code: { hidden: false } },
   },
+  decorators: [
+    // Workaround Storybook's wrapping flex parent that make everything full width
+    story => <div>{story()}</div>,
+  ],
 } as ComponentMeta<typeof Combobox>;
+
+const BasicCombobox: ComponentStory<typeof Combobox> = args => {
+  const [selected, setSelected] = useState<ComboboxOption[]>([]);
+
+  return (
+    <Combobox {...args} heading="Teammates">
+      <Combobox.Content
+        options={[
+          { id: "1", label: "Bilbo Baggins" },
+          { id: "2", label: "Frodo Baggins" },
+          { id: "3", label: "Pippin Took" },
+          { id: "4", label: "Merry Brandybuck" },
+          { id: "5", label: "Sam Gamgee" },
+        ]}
+        onSelect={selection => {
+          setSelected(selection);
+        }}
+        selected={selected}
+      >
+        <Combobox.Action
+          label="Add Teammate"
+          onClick={() => {
+            alert("Added a new teammate ✅");
+          }}
+        />
+        <Combobox.Action
+          label="Manage Teammates"
+          onClick={() => {
+            alert("Managed teammates 👍");
+          }}
+        />
+      </Combobox.Content>
+    </Combobox>
+  );
+};
 
 const ComboboxButton: ComponentStory<typeof Combobox> = args => {
   const [selected, setSelected] = useState<ComboboxOption[]>([]);
 
   return (
     <Combobox {...args}>
-      <Combobox.TriggerButton
-        label="Select Teammate"
-        variation="subtle"
-        type="primary"
-        icon="arrowDown"
-        iconOnRight={true}
-      />
+      <Combobox.Activator>
+        <Button
+          label="Select Teammate"
+          variation="subtle"
+          type="primary"
+          icon="arrowDown"
+          iconOnRight={true}
+        />
+      </Combobox.Activator>
       <Combobox.Content
         options={[
           { id: "1", label: "Bilbo Baggins" },
@@ -40,7 +82,7 @@ const ComboboxButton: ComponentStory<typeof Combobox> = args => {
           { id: "12", label: "Samwise Gamgee" },
           { id: "14", label: "Faramir" },
         ]}
-        onSelect={selection => {
+        onClose={selection => {
           setSelected(selection);
         }}
         selected={selected}
@@ -75,7 +117,9 @@ const ComboboxChip: ComponentStory<typeof Combobox> = args => {
       selected={selected}
       subjectNoun="teammates"
     >
-      <Combobox.TriggerChip label="Teammates" />
+      <Combobox.Activator>
+        <Chip variation="subtle" label="Teammates" />
+      </Combobox.Activator>
       <Combobox.Option id="1" label="Bilbo Baggins" />
       <Combobox.Option id="2" label="Frodo Baggins" />
       <Combobox.Option id="3" label="Pippin Took" />
@@ -103,14 +147,7 @@ const ComboboxEmptyState: ComponentStory<typeof Combobox> = args => {
   const [selected, setSelected] = useState<ComboboxOption[]>([]);
 
   return (
-    <Combobox {...args}>
-      <Combobox.TriggerButton
-        label="Select Teammate"
-        variation="subtle"
-        type="primary"
-        icon="arrowDown"
-        iconOnRight={true}
-      />
+    <Combobox {...args} heading="Teammates">
       <Combobox.Content
         options={[]}
         onSelect={selection => {
@@ -139,20 +176,18 @@ const ComboboxClearSelection: ComponentStory<typeof Combobox> = args => {
   ]);
 
   return (
-    <>
+    <div
+      style={{
+        display: "flex",
+        flexDirection: "row",
+      }}
+    >
       <ClearButton
         label="Clear Selection"
         type="primary"
         onClick={() => setSelected([])}
       />
-      <Combobox {...args}>
-        <Combobox.TriggerButton
-          label="Select a teammate"
-          variation="subtle"
-          type="primary"
-          icon="arrowDown"
-          iconOnRight={true}
-        />
+      <Combobox {...args} heading="Teammates">
         <Combobox.Content
           options={[
             { id: "1", label: "Bilbo Baggins" },
@@ -189,7 +224,7 @@ const ComboboxClearSelection: ComponentStory<typeof Combobox> = args => {
           />
         </Combobox.Content>
       </Combobox>
-    </>
+    </div>
   );
 };
 
@@ -199,6 +234,7 @@ const ComboboxMultiSelection: ComponentStory<typeof Combobox> = args => {
   return (
     <Combobox
       multiSelect
+      heading="Teammates"
       onSelect={selection => {
         setSelected(selection);
       }}
@@ -250,6 +286,9 @@ const ComboboxMultiSelection: ComponentStory<typeof Combobox> = args => {
     </Combobox>
   );
 };
+
+export const Basic = BasicCombobox.bind({});
+Basic.args = {};
 
 export const TriggerButton = ComboboxButton.bind({});
 TriggerButton.args = {};