+++ packages/components/src/Combobox/Combobox.test.tsx
@@ -3,6 +3,7 @@ import React from "react";
 import { ComboboxOption } from "./Combobox.types";
 import { Combobox } from "./Combobox";
 import { COMBOBOX_TRIGGER_COUNT_ERROR_MESSAGE } from "./hooks/useComboboxValidation";
+import { Chip } from "../Chip";
 
 // jsdom is missing this implementation
 const scrollIntoViewMock = jest.fn();
@@ -26,11 +27,8 @@ describe("Combobox validation", () => {
     }).not.toThrow();
   });
 
-  it("throws an error if there is no Trigger element", () => {
-    expect.assertions(1);
-    let error;
-
-    try {
+  it("renders without error when there is no TriggerButton or TriggerChip", () => {
+    expect(() => {
       render(
         <Combobox>
           <Combobox.Content
@@ -40,11 +38,24 @@ describe("Combobox validation", () => {
           ></Combobox.Content>
         </Combobox>,
       );
-    } catch (e) {
-      error = e as Error;
-    } finally {
-      expect(error?.message).toBe(COMBOBOX_TRIGGER_COUNT_ERROR_MESSAGE);
-    }
+    }).not.toThrow();
+  });
+
+  it("renders without error when there is a ComboboxActivator", () => {
+    expect(() => {
+      render(
+        <Combobox>
+          <Combobox.Activator>
+            <Chip variation="subtle" label="Teammates" />
+          </Combobox.Activator>
+          <Combobox.Content
+            options={[]}
+            onSelect={jest.fn()}
+            selected={[]}
+          ></Combobox.Content>
+        </Combobox>,
+      );
+    }).not.toThrow();
   });
 
   it("throws an error if there are multiple of the same Trigger element", () => {
@@ -91,6 +102,25 @@ describe("Combobox validation", () => {
     } finally {
       expect(error?.message).toBe(COMBOBOX_TRIGGER_COUNT_ERROR_MESSAGE);
     }
+  });
+
+  it("throws an error if there is a Trigger element and a ComboboxActivator", () => {
+    const component = () =>
+      render(
+        <Combobox>
+          <Combobox.TriggerButton label="Button" />
+          <Combobox.Activator>
+            <Chip variation="subtle" label="Teammates" />
+          </Combobox.Activator>
+          <Combobox.Content
+            options={[]}
+            onSelect={jest.fn()}
+            selected={[]}
+          ></Combobox.Content>
+        </Combobox>,
+      );
+
+    expect(component).toThrow(COMBOBOX_TRIGGER_COUNT_ERROR_MESSAGE);
   });
 
   it("throws an error if there are multiple Trigger elements and no Content", () => {