+++ packages/components/src/Combobox/components/ComboboxTrigger/index.ts
@@ -1,2 +1,3 @@
 export { ComboboxTriggerButton } from "./ComboboxTriggerButton/ComboboxTriggerButton";
 export { ComboboxTriggerChip } from "./ComboboxTriggerChip/ComboboxTriggerChip";
+export { ComboboxTrigger } from "./ComboboxTrigger";